Fix tensor shape incompatibilities
Identity didn't change depth of the tensor, so a padding in depth is done if necessary so that addition can be performed properly at the end of the block.
Also, perform a pointwise convolution at the end of the cell to reduce the depth of concatenation tensor to the filters value, so that is not increased (PNAS increase it only in reduction cell).