Add computation of partition sizes between cells of a CNN
Partitions can be useful for deploying a big CNN into the edge. To partition properly the net, the size of the tensors to pass between each partition must be computed, so that it's possible to cut the CNN where the size is optimal. Partition size info is stored in partitions.txt file under each CNN folder (see tensorflow_cnn log folder structure).