Add training_accuracy.csv, add data_augmented field to csv files
The training_accuracy.csv file can be used for training ML regressors also for the accuracy predictions. The "data_augmented" field instead is a boolean that states if the record has been generated through data augmentation or if it is "original". This field is excluded from training for any regressor (through their respective configs), but could be used for filtering the dataset if necessary.