Add Predictor class hierarchy
Predictor is an abstract class, used to standardize the interface of all methods used in POPNAS for regression of time and accuracy. It's still a WIP, but allows to use in a similar way all techniques based ML and NN tested so far.

regressor_testing_v2.py provides an example on how to use the new interface for testing the predictor on data of a previous run.