Update and improve predictors hierarchy, add Conv1D predictor
Add NNPredictor as intermediate abstract class for NN implemented with Keras, where train and predict procedures are defined with keras but new private abstract methods for building models and dataset are required. To build different NN predictors is so required to just specify how to generate the Keras model and Tensorflow dataset.

Now aMLLibrary 'y' config key is specified in class for more flexibility, instead of hardcoding it in config file.