Remove code for 'run restore' command line option
This code was faulty with actual implementation and it was also probably faulty on v1, so it is directly removed and a NotImplementedError is raised in case the restore parameter is provided. Instead of using 3 different restore parameters, only the log folder is now required, since it is better to leave the details to the implementation logic, when it will be actually implemented (less error-prone).