Add exploration step to POPNAS algorithm
Exploration step is used to train relevant networks that use inputs and operators that are never or very rarely used inside the pareto front. This allows to explore and know better operation and inputs with fewer samples, that can be still relevant in later stages of the progressive neural architecture search, even if they performed bad in previous steps.
These exploration networks are also added to the networks to be expanded in next stage, so that if they actually perform well they can also be further utilized.