Improve training procedure for children CNNs
Now the learning rate is adapted during the epochs by using the cosine decay scheduler, this change seems to benefit a lot the convergence and best validation accuracy found. SGD seems to perform worse than ADAM on 20 epochs, but could perform better for final model training script. It's still unclear if CosineDecayRestart is generally better than CosineDecay, but from small tests it seems so and for now it's used as the learning rate schedule for POPNAS run.
Also note that learning rate has been changed from 0.01 to 0.05.