Add ScheduledDropPath to CNNs
ScheduledDropPath can be enabled in json config, by setting a value between 0 and 1 to "drop_path_prob". 0 means that ScheduledDropPath will not be used.
ScheduledDropPath implementation is inspired from Tensorflow repository, but it is actually a bad implementation, since it doesn't ensure that at least a path survives. This is probably the cause that leads to awful performance if ScheduledDropPath is enabled, for now it is better to disable it by setting it to 0 and investigate a solution later.