Refactor Predictors hierarchy to allow hyperparams search
nn_predictor has been renamed to keras_predictor and adapted to allow training with a set of defined parameters (like before) or using an external library for tuning automatically the hyperparameters. Validation datasets can now be generated for the Keras models, since validation loss must be used for proper tuning.

Right now Ray Tune is used to search the hyperparameters, but it terminates with an error when tune.run is executed.