Change KerasPredictors' datasets batch size to 4
Before it was set to 1, since dataset is quite small, but can grow to 1k samples in last step becoming very slow. Especially for hp tuning is important to reduce the training time to an acceptable target, since multiple networks must be trained. 4 seems a good compromise between total training steps and total training time, but could be increased more if necessary.