Solve problem of multi-branch model when using multi-output
This change involved a major change in ModelGenerator logic, that simplifies how multiple things are handled. Now the logic precomputes where the input normalization must be applied and which cells are actually used in the model (when input -1 is not used only some cells are useful). These changes allow to not build the cell entirely if not needed and gracefully handle all problems that arise from the change, simplifying also the partitioning and improving the general performances of the model generation.