Improve catboost parameters
These parameters seems to lead to a bit of error reduction on average, in pessimistic cases they seems to achieve same errors on some B and still improve the others. The tests have been performed on 3 different runs with 3 different configurations and GPU hardware, so the configuration seems quite robust.