Remove identity and pool operations since they are bugged
Since they don't reshape depth, multiple concatenations multiplies progressively the depth of the input tensor. This exponential increase in tensor depth cause the allocation of unnameable amounts of memory, to which the GPU just says f*** this and crash all this mess with an OOM error. Therefore, don't use them.