Update best model script to generalize on any model and dataset
The script has been renamed to final_training.py. It now handles also multi-output models and can adapt the training configuration through an external JSON file, like the search algorithm. In this way, it's possible to use different datasets and also to adapt the model parameters to perform many types of tests, like increasing the cells and the filters, in a much easier way.