Make saving ONNX optional
A boolean parameter can be set in the json config to enable the saving of the ONNX file. These files can waste a lot of space on disk but do not require much time to generate, so it is mainly useful to disable ONNX conversion if you want to minimize the disk space required.