Add GraphGenerator
It can be used as the counterpart of ModelGenerator, used to build only the network graph. It is used mainly to retrieve the number of parameter of a network, without actually building it in Keras (it is more than 100x faster in this way). Also, the graph can be useful for external analysis.