Adapt search algorithm to support different device train strategies
The goal of this commit and further commit adjustments is to support POPNAS execution on different devices (CPU, GPU, TPU) and possibly on distributed environments (multiple workers or multiple devices in a single worker), by exploiting the strategies already defined in Tensorflow.

When complete, the code should run on any device without further modifications, only the configuration file fields regarding the strategy should be set properly.