Add config parameter for specifying the objectives of Pareto optimization
The run will adapt based on the selected objectives, training only the required predictors and adapting the domination rule. Accuracy metric must always be specified, otherwise the run would be meaningless (select identity and pooling only since they have the least amount of parameters and are very fast to train).