Add a "fake" training to avoid dataset lazy initialization time bias
Since dataset is lazy loaded, for cases where a custom dataset loaded from images on disks, the training time of the first epoch of the first network incurred an extremely large overhead. This change promotes a more fair evaluation of the training time of the first network trained in the session (empty cell as common start, any cell on restore).