Add dataset generator for time series classification
For now, only custom datasets from .npy files are supported, providing only the training set which can be stratified split into validation, as usual.
Support for custom validation and test sets, plus other formats will come in future commits, this is a bare minimum for integrating it in POPNAS and work on actual NN models.