Update ModelGenerator to support time series
Operator instantiation is managed through the usage of a new class, defined in ops module.
This class helps in adapting the operators of the model, based on the input shape (e.g. use 2D layers for images and 1D for time series).
Custom Keras layers and ModelGenerator logic has been updated to support the new input shape case.