Add weight regularization to CvT and parameter for head dimension
Also perform some refactorings, in particular bloatware code has been added to implement get_config for each subclass of CvT, so that they can be serialized better (model save). A custom dropout is also applied, which become an identity if rate = 0, maybe improving the performances a bit if not optimized directly in tensorflow.