Add NATS-bench proxy modules, to integrate in POPNAS
Benchmarking module provide the interface for querying NATS-bench topology search space given a POPNAS cell.
NetworkBenchManager provide the same public interface of NetworkManager, but instead of producing the models and training them it simply queries NATS-bench for results.
These modules must still be integrated in POPNAS algorithm, when the benchmark option is set.