Tentative fix for Tensorflow memory leaks
Memory leaks are a problem in E2E scripts, especially the TSC one which runs multiple experiments. This commit uniform the way for garbage collecting and execute it after each network training. Hopefully it will tone down the memory footprint a bit. Multiple issues are present in the Tensorflow repository, but the causes of the leaks are unclear.