Add config option for using Squeeze-Excitation on cell output
SE is totally optional and when output is also residual, the SE block is placed before the residual sum, following the SE paper. By default, SE is disabled in the search configurations, since it cause a significant time overhead in architectures with 1 block and boost more operators which do not have good performances, which could cause the accuracy predictor to think that these operators are good while is the SE block doing the heavy work. Just a thought, extended experiments are required.