Extend operator layers to allow setting different activation functions
By default, the activation function is still set to Swish, but it should be easier to change the activation when necessary. The allocators have not been extended yet since the operators regexes do not expose a parameter for the activation function, but it is technically possible to extend them in the future. This change is more aimed for cases external to the operators defined in the search space.