Fix tensorboard network ID on run restore
The network ID was always initialized to 0, meaning that the training data of a network could override previous ones if the folder already existed (when training B=1 networks, since the folders are separated based on the number of blocks).
Does not change the restore file, just check the training results csv to infer the number of networks.