Add graph generator instance to model generator
Graphs could be seen as proxy generation of actual networks, so it makes sense to have the graph generators inside the model generators. DAGs should follow the same structure of the model generator macro-architecture, but right now it's kept as a single class (careful when using params in the Pareto front).