Add PredictorsHandler utility class
This class can be used during the search to initialize the predictors and generate the features for them in an easy way. Functions from feature_utils or direct instantiation of the predictors are still possible when needed, the class just acts as a "shortcut" for keeping the references and providing a simplified interface. Logic for class __init__ is similar to the one previously used in Popnas class.