Add CellSpecification and BlockSpecification
These classes are specific for the POPNAS architecture genotype, getting rid of the "list of tuples" generic structure present everywhere.
Now they are standardized, and they come with multiple utility functions which should help in making the code more readable and less verbose.
A good refactor has been done in this commit, and even if it runs fine, it could have some silent bugs, since this commit is basically a tactical nuke auto-fired inside the codebase (touching almost all files of the codebase) :)