Update ModelGenerator to use WrappedTensor new class
WrappedTensors are necessary to allow undefined input sizes for the spatial resolution, so that fully CNN can use different image sizes.
This class stores the shape of the tensor together with the tensor itself, but this shape can be stored as a ratio of original input size, making it a flexible tool to address tensor shape conversions without fixing a specific size.