Modify the SegmentationModelGenerator macro-architecture design
The upsample units have been modified to process the lookbacks with the usual position, -1 with bilinear up - concat - 1x1 conv, -2 with transpose conv.
The upsample unit is now generated as part of an upsample cell, so if this cell is skipped also the upsample unit will not be generated. "Skip lookbacks" can still be upsampled with transpose convolutions when necessary.