Refactor datasets pipeline to use Albumentations
Unfortunately, the preprocessing and augmentation functions can't be vectorized by TF dataset (numpy functions are not executed in parallel either), so even if Albumentations transformations are fast on a single image, they perform worse compared to Keras ones which act directly on tensors.
The API is also changed, could be interesting to test other types of data augmentation libraries.