Add build_model_graph to BaseModelGenerator
This function will replace the GraphGenerator, directly mapping the functionality into the ModelGenerators. Since the macro structure is defined by models actual implementation, also the network graph must follow the same structure, therefore is better to have them together in the same class, sharing the parameters (in GraphGenerator version there is a lot of parameters passing and duplication).