Conciliate segmentation sample weights with balance_class_weights flag
Basically, now sample weights are used only in segmentation tasks when balance_class_weights is set to true in the configuration.
The dataset class changes the flag with a side effect, so that the class weights are not applied in the Keras function, which can not use them on image labels.