Make DeepLab-like empty cell similar to other networks
Before this change, the empty cell used ASPP on the whole image and applied no decoder, so the time required for training was not indicative of the shared layers of all architecture. Still, this implementation adds some overhead since it uses avgpool and pointwise conv to adapt the shape, but compared to ASPP and decoder this overhead is negligible.