Solve mixed precision problem with predictors by fixing dtype to FP64
Keras predictors should now fix the dtype of each layer to self.dtype, which defaults to float64 (defined in Keras predictor). This solves issues when using mixed precision, both regarding stability and type mismatches. FP64 seems more performant and stable than FP32, especially when predicted values are close to each other, and the performance cost is negligible in this case.