Deprecate perform_on_gpu data augmentation option
Right now, Keras augmentation layers are slow as snails both on CPU and GPU. Furthermore, the actual implementation in POPNAS worked correctly only for image classification tasks. Since a refactor was already needed, I decided to simply remove the option, given the fact that it is almost always better to execute the data augmentation asynchronously on the CPU. ModelGenerators still have the logic to support Keras data augmentation models, if they will be needed in the future.