Move CellSpecification and BlockSpecification into a separate module
Having them into the search_space module caused a circular dependency in save cell image script, and could have been problematic in the future. CellSpecification is imported for type notation in A LOT of modules (see changes!), so this is a change to prevent potential circular dependencies in the future.