Add inference predictor logic to procedure and controller
New files are produced in a similar way to training time, storing the features for inference time and the dynamic reindex computed for inference. The predictor is an SVR like the training time predictor, but will be trained separately on the inference features.

If inference time is not provided to the config, the inference predictor will not be instantiated.