Fix ZoomedConvolutions for inputs with flexible resolution
Using the shape provided by build function was slightly more efficient, but the resolution dimensions resulted to None when the input shape is not fixed (e.g., validation performed on images with different shapes). Now the shape is directly retrieved in call function, and it seems to not slow down the layer execution significantly.